+++ CHANGELOG.md
@@ -5,10 +5,24 @@ All notable changes to this project will be documented in this file.
 ## [0.5.0] - 2021-04-08
 
 ### New
+- `clang` controller for C++ compiler:   
+  install, compile with 1 command, manage versions, etc.   
+- `signer` controller for signer registry management:   
+  generate keys, import keys and mnemonics, specify default signer, etc.   
+- `network` controller for network registry management:   
+  add networks, configure network giver, specify default network, etc.   
+- `contract` controller for contract management:   
+- get contract info summary, top up balance, deploy, run, run-local, run-executor   
+- `tondev js wrap` generates `code` and `codeHash` fields in addition to `tvc`.   
+- `tondev contract info` prints code hash.   
+- `tondev info` command prints a summary information from all controllers.   
+- short aliases for controllers, commands and options.   
+  For example instead of using `tondev network list`    
+  you can use `tondev n l` and even shorter `tondev nl`.   
 
-- `signers` controller for signer registry management. 
-- `networks` controller for network registry management. 
-- `contract` controller for contract management. 
+### Fixed
+
+- `tvm_linker` downloads always even if it's up to date.
 
 ## [0.4.0] - 2021-03-25